requirements.txt
fastapi
uvicorn[standard]
python-dotenv
groq
httpx
requests
fpdf2
